IPython Console: Move spyder_kernel.py to utils/ipython
