+++ spyder/utils/ipython/spyder_kernel.py
